requirements.txt
openai>=1.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0
flask>=2.3.0
python-dotenv>=1.0.0
flask-cors>=3.0.0